Properly separate event sink and dispatcher
